renamed top level dir to manipulation
